% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inexact_join.R
\name{inexact_join}
\alias{inexact_join}
\title{Performs a fuzzy (inexact) join, trusting the algorithm by default.}
\usage{
inexact_join(x, y, by, max_dist = Inf, method = c("osa", "lv", "dl",
  "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex"),
  mode = "inner", custom_match = NULL, ignore_case = FALSE,
  match_cols = FALSE, ...)
}
\description{
The function allows to provide custom matches.
}
