% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trust_join.R
\name{trust_join}
\alias{trust_join}
\title{Perform a fuzzy join trusting the algorithm.}
\usage{
trust_join(x, y, by, max_dist = Inf, method = c("osa", "lv", "dl",
  "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex"),
  mode = "inner", custom_match = NULL, ignore_case = FALSE,
  match_cols = FALSE, ...)
}
\description{
Performs a fuzzy join that trusts the algorithm. Allows to provide custom matches.
}
